# [android] use Type.GetType() to load AppLinks