@PureWeen 2019-12-20
One UI test failing on API 19 that's unrelated
- CarouselViewHorizontal("CarouselView (Code, Horizontal)")_google_pixel_2-8_1_0